Markdown Editor added, other UI changes

+++ src/app/profile/page.tsx
@@ -5,13 +5,15 @@ import { BlogPost } from "@/types/types";
 import { DeleteBlog } from "@/lib/client";
 import { useAccount } from 'wagmi'
 import Link from 'next/link';
+import MarkdownIt from 'markdown-it';
+
+const md = new MarkdownIt();
 
 const Profile = () => {
     const [blogPosts, setBlogPosts] = useState<readonly BlogPost[]>([]);
 
     const { address } = useAccount()
 
-console.log("addaa ",address)
     const loadBlogs = async () => {
         if (typeof window !== "undefined") {
             const allBlogs = await GetAllBlogs(window.ethereum);
@@ -28,13 +30,19 @@ console.log("addaa ",address)
     }, []);
 
     const handleDelete = async (blogId: bigint) => {
-        try {
+        const confirmed = window.confirm("Are you sure you want to delete this blog?");
+      
+        if (confirmed) {
+          try {
             await DeleteBlog(window.ethereum, blogId);
             // setBlogPosts(blogPosts.filter((blog) => blog.id !== BigInt(blogId)));
-        } catch (error) {
+          } catch (error) {
             console.error("Error deleting blog post:", error);
+          }
+        } else {
+          console.log("Deletion cancelled");
         }
-    };
+      };
 
    
     return (
@@ -56,13 +64,10 @@ console.log("addaa ",address)
                                 <Link
                                 href={`/blog/${post.id}`} 
                                     key={post.id}
-                                    className="bg-gray-950  p-5 m-5 hover:border-3 hover:shadow-md hover:shadow-white hover:border-teal-400 border-gray-400 border-2 rounded-xl shadow-md"
+                                    className="bg-gray-950 lg:min-h-60 p-5 m-5 hover:border-3 hover:shadow-md hover:shadow-white hover:border-teal-400 border-gray-400 border-2 rounded-xl shadow-md"
                                 >
                                     <h2 className="text-white text-xl font-bold mb-2">{post.title}</h2>
-                                    <p className="text-gray-300">
-                                        {post.content.slice(0, 100)}
-                                        {post.content.length > 30 && "..."}
-                                    </p>
+                                    <p className="text-gray-300" dangerouslySetInnerHTML={{ __html: post.content.length>30? md.render(post.content).slice(0, 100)+"..." :   md.render(post.content) }} />
                                     <p className="text-gray-400 pt-5">Author: {post.author.slice(0, 20)}</p>
                                     <button
                                         className="bg-red-500 hover:bg-red-700 hover:shadow-md hover:shadow-red-400 mt-4 text-white py-2 px-4 rounded focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-700"
